+++ Private/Get-DeploymentTypeInfo.ps1
@@ -1,6 +1,7 @@
 <#
 .Synopsis
 Created on:   28/10/2023
+Update on:    23/03/2024
 Created by:   Ben Whitmore
 Filename:     Get-DeploymentTypeInfo.ps1
 
@@ -25,6 +26,13 @@ function Get-DeploymentTypeInfo {
 
         # Create an empty array to store the deployment type information
         $deploymentTypes = @()
+
+        # Characters that are not allowed in Windows folder names
+        $invalidChars = '[<>:"/\\|\?\*]'
+
+        # Get UTF-8 encoding without BOM
+        $utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $false 
+        
     }
     process {
 
@@ -58,6 +66,27 @@ function Get-DeploymentTypeInfo {
                         $uninstallLocation = $object.Installer.Contents.Content.Location
                     }
 
+                    # Sanitize the folder names
+                    $applicationNameSanitized = ($xmlContent.AppMgmtDigest.Application.title.'#text' -replace $invalidChars, '_').TrimEnd('.', ' ')
+                    $deploymentTypeNameSanitized = ($object.Title.InnerText -replace $invalidChars, '_').TrimEnd('.', ' ')
+
+                    # Detection Methods child folder path
+                    $detectionMethodsFolderPath = ("{0}\{1}" -f $applicationNameSanitized, $deploymentTypeNameSanitized)
+
+                    # Build final folder name strings
+                    $detectionMethodsFolder = Join-Path -Path "$workingFolder_Root\DetectionMethods" -ChildPath $detectionMethodsFolderPath
+
+                    # Create the Detection Methods folder
+                    New-FolderToCreate -Root "$workingFolder_Root\DetectionMethods" -FolderNames $detectionMethodsFolderPath
+
+                    # Remove existing files from the Detection Methods folder to avoid ambiguity on import
+                    $existingFiles = [System.IO.Directory]::GetFiles($detectionMethodsFolder)
+                    
+                    foreach ($file in $existingFiles) {
+                        Write-Log -Message ("Removing existing file '{0}'" -f $file) -LogId $LogId -Severity 2
+                        [System.IO.File]::Delete($file) 
+                    }
+
                     # Create a new custom hashtable to store Deployment type details
                     $deploymentObject = [PSCustomObject]@{}
 
@@ -65,20 +94,166 @@ function Get-DeploymentTypeInfo {
                     $deploymentObject | Add-Member NoteProperty -Name Application_Id -Value $ApplicationId
                     $deploymentObject | Add-Member NoteProperty -Name ApplicationName -Value $xmlContent.AppMgmtDigest.Application.title.'#text'
                     $deploymentObject | Add-Member NoteProperty -Name Application_LogicalName -Value $xmlContent.AppMgmtDigest.Application.LogicalName
-                    $deploymentObject | Add-Member NoteProperty -Name LogicalName -Value $Object.LogicalName
-                    $deploymentObject | Add-Member NoteProperty -Name Name -Value $Object.Title.InnerText
-                    $deploymentObject | Add-Member NoteProperty -Name Technology -Value $Object.Installer.Technology
-                    $deploymentObject | Add-Member NoteProperty -Name ExecutionContext -Value $Object.Installer.ExecutionContext
+                    $deploymentObject | Add-Member NoteProperty -Name LogicalName -Value $object.LogicalName
+                    $deploymentObject | Add-Member NoteProperty -Name Name -Value $object.Title.InnerText
+                    $deploymentObject | Add-Member NoteProperty -Name Technology -Value $object.Installer.Technology
+                    $deploymentObject | Add-Member NoteProperty -Name ExecutionContext -Value $object.Installer.ExecutionContext
                     $deploymentObject | Add-Member NoteProperty -Name InstallContent -Value $installLocation.TrimEnd('\') 
-                    $deploymentObject | Add-Member NoteProperty -Name InstallCommandLine -Value $Object.Installer.CustomData.InstallCommandLine
-                    $deploymentObject | Add-Member NoteProperty -Name UnInstallSetting -Value $Object.Installer.CustomData.UnInstallSetting
+                    $deploymentObject | Add-Member NoteProperty -Name InstallCommandLine -Value $object.Installer.CustomData.InstallCommandLine
+                    $deploymentObject | Add-Member NoteProperty -Name UnInstallSetting -Value $object.Installer.CustomData.UnInstallSetting
                     $deploymentObject | Add-Member NoteProperty -Name UninstallContent -Value $uninstallLocation.TrimEnd('\') 
-                    $deploymentObject | Add-Member NoteProperty -Name UninstallCommandLine -Value $Object.Installer.CustomData.UninstallCommandLine
-                    $deploymentObject | Add-Member NoteProperty -Name ExecuteTime -Value $Object.Installer.CustomData.ExecuteTime
-                    $deploymentObject | Add-Member NoteProperty -Name MaxExecuteTime -Value $Object.Installer.CustomData.MaxExecuteTime
-                
-                    Write-Log -Message ("Application_Id = '{0}', Application_Name = '{1}', Application_LogicalName = '{2}', LogicalName = '{3}', Name = '{4}', Technology = '{5}', ExecutionContext = '{6}', InstallContext = '{7}', `
-                 InstallCommandLine = '{8}', UninstallSetting = '{9}', UninstallContent = '{10}', UninstallCommandLine = '{11}', ExecuteTime = '{12}', MaxExecuteTime = '{13}'" -f `
+                    $deploymentObject | Add-Member NoteProperty -Name UninstallCommandLine -Value $object.Installer.CustomData.UninstallCommandLine
+                    $deploymentObject | Add-Member NoteProperty -Name ExecuteTime -Value $object.Installer.CustomData.ExecuteTime
+                    $deploymentObject | Add-Member NoteProperty -Name MaxExecuteTime -Value $object.Installer.CustomData.MaxExecuteTime
+                    $deploymentObject | Add-Member NoteProperty -Name DetectionProvider -Value $object.Installer.DetectAction.Provider
+
+                    # Switch on the detection method and save to file
+                    Switch ($object.Installer.DetectAction.Provider) {
+
+                        # If Detection Method is a 'Script'
+                        'Script' {
+                            $detectionTypeScriptBody = $object.Installer.DetectAction.Args.Arg.Where({ $_.Name -eq 'ScriptBody' }).InnerText
+                            $detectionTypeScriptRunAs32Bit = $object.Installer.DetectAction.Args.Arg.Where({ $_.Name -eq 'RunAs32Bit' }).InnerText
+                            $detectionTypeExecutionContext = $object.Installer.DetectAction.Args.Arg.Where({ $_.Name -eq 'ExecutionContext' }).InnerText
+                            $detectionTypeScriptType = $object.Installer.DetectAction.Args.Arg.Where({ $_.Name -eq 'ScriptType' }).InnerText
+        
+                            # ScriptType
+                            # 0 = PowerShell
+                            # 1 = VBScript
+                            # 2 = JavaScript
+        
+                            Switch ($detectionTypeScriptType) {
+                                '0' {
+                                    $detectionTypeScriptFileExtension = '.ps1'
+                                }
+                                '1' {
+                                    $detectionTypeScriptFileExtension = '.vbs'
+                                }
+                                '2' {
+                                    $detectionTypeScriptFileExtension = '.js'
+                                }
+                            }
+        
+                            # Extract only the encoded base64 script from the script body
+                            $pattern = '# ENCODEDSCRIPT # Begin Configuration Manager encoded script block #\s*(.*?)\s*# ENCODEDSCRIPT# End Configuration Manager encoded script block'
+                            $matchSigned = [regex]::Match($detectionTypeScriptBody, $pattern, [System.Text.RegularExpressions.RegexOptions]::Singleline)
+        
+                            if ($matchSigned.Success) {
+                                $extractedContent = $matchSigned.Groups[1].Value
+
+                                # Decode the Base64 string
+                                #$Encoding = [System.Text.Encoding]::GetEncoding('windows-1252')
+                                #$scriptContent = $Encoding.GetString([System.Convert]::FromBase64String($extractedContent))
+                                $scriptContent = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($extractedContent))
+
+                                # We need to trim the script to honor digital signatures
+                                # Split the original string into lines
+                                $lines = $scriptContent -split "`n"
+
+                                # Remove the last line because the xml adds padding before the encoded string
+                                $lines = $lines[0..($lines.Count - 2)]
+
+                                # Join the lines back into a single string
+                                $finalScriptContent = $lines -join "`n"
+                            }
+                            else {
+                                
+                                # No need to deal with encoding, just pass the script body to the final object for saving
+                                $finalScriptContent = $detectionTypeScriptBody
+                            }
+  
+                            # Write the detection method to a file
+                            $detectionMethodFile = Join-Path -Path $detectionMethodsFolder -ChildPath "DetectionScript$detectionTypeScriptFileExtension"
+   
+                            try {
+
+                                # Specifying Out-File with Encoding UTF8 is not honored so use .NET method instead to write to file
+                                [System.IO.File]::WriteAllText($detectionMethodFile, $finalScriptContent, $utf8NoBomEncoding)
+                                Write-Log -Message ("Detection method script saved to file '{0}'" -f $detectionMethodFile) -LogId $LogId
+                                Write-Host ("Detection method script saved to file '{0}'" -f $detectionMethodFile) -ForegroundColor Green
+                            }
+                            catch {
+                                Write-Log -Message ("Could not write detection method to file '{0}'" -f $detectionMethodFile) -LogId $LogId -Severity 3
+                                Write-Host ("Could not write detection method to file '{0}'" -f $detectionMethodFile) -ForegroundColor Red
+                            }
+                        }
+
+                        # If Detection Method is File/Reg/MSICode
+                        'Local' {
+                            $detectionTypeExecutionContext = $object.Installer.DetectAction.Args.Arg.Where({ $_.Name -eq 'ExecutionContext' }).InnerText
+                            $detectionTypeMethodBody = $object.Installer.DetectAction.Args.Arg.Where({ $_.Name -eq 'MethodBody' }).InnerText
+                            
+                            # Define the detection method Xml file path
+                            $detectionMethodXmlFile = Join-Path -Path $detectionMethodsFolder -ChildPath 'MethodBody.xml'
+                        
+                            # Write the detection method to an Xml file
+
+                            try {
+                                # Specifying Out-File with Encoding UTF8 is not honored so use .NET method instead
+                                [System.IO.File]::WriteAllText($detectionMethodXmlFile, $detectionTypeMethodBody, $utf8NoBomEncoding)
+                                Write-Log -Message ("Detection method XML saved to file '{0}'" -f $detectionMethodXmlFile) -LogId $LogId
+                                Write-Host ("`Detection method XML saved to file '{0}'" -f $detectionMethodXmlFile) -ForegroundColor Cyan
+                            }
+                            catch {
+                                Write-Log -Message ("Could not write detection method to file '{0}'" -f $detectionMethodXmlFile) -LogId $LogId -Severity 3
+                                Write-Host ("Could not write detection method to file '{0}'" -f $detectionMethodXmlFile) -ForegroundColor Red
+                            }
+
+                            # Attempt to extract the local detection methods from the XML. We will ignore any 'or' operators as these are not supported in Intune
+                            if (Test-Path -Path $detectionMethodXmlFile) {
+                                $localDetectionMethods = Get-DetectionMethod -LogId $LogId -XMLObject $detectionTypeMethodBody
+                                
+                                if ($localDetectionMethods.Count -gt 0) {
+
+                                    Write-Log -Message 'Local Detection Methods extracted from XML' -LogId $LogId
+                                    Write-Host "`nLocal Detection Methods extracted from XML" -ForegroundColor Cyan
+
+                                    foreach ($method in $localDetectionMethods) {
+                                        Write-Log -Message ("{0}" -f $method) -LogId $LogId
+                                        Write-Host ("{0}" -f $method) -ForegroundColor Green
+                                    }
+                                    
+                                    # Export the detection method to a Json file
+                                    # Define the detection method Json file path
+                                    $detectionMethodJsonFile = Join-Path -Path $detectionMethodsFolder -ChildPath 'DetectionMethod.json'
+
+                                    # Create a detection method json file
+                                    $detectionJson = New-IntuneDetectionMethod -LocalSettings $localDetectionMethods
+
+                                    try {
+                                        # Specifying Out-File with Encoding UTF8 is not honored so use .NET method instead
+                                        [System.IO.File]::WriteAllText($detectionMethodJsonFile, $detectionJson, $utf8NoBomEncoding)
+                                        Write-Log -Message ("Intune detection method Json saved to file '{0}'" -f $detectionMethodJsonFile) -LogId $LogId
+                                        Write-Host ("`Intune detection method Json saved to file '{0}'" -f $detectionMethodJsonFile) -ForegroundColor Cyan
+                                    }
+                                    catch {
+                                        Write-Log -Message ("Could not create Intune detection method Json file '{0}'" -f $detectionMethodJsonFile) -LogId $LogId -Severity 3
+                                        Write-Host ("Could not create Intune detection method Json file '{0}'" -f $detectionMethodJsonFile) -ForegroundColor Red
+                                    }
+                                }
+                                else {
+                                    Write-Log -Message ("There was an error getting the local detection methods for deployment type '{0}' from the XML" -f $detectionMethodXmlFile) -LogId $LogId -Severity 3
+                                    Write-Host ("There was an error getting the local detection methods for deployment type '{0}' from the XML" -f $detectionMethodXmlFile) -ForegroundColor Red
+                                }
+                            }
+                            else {
+                                Write-Log -Message ("There was an error getting the local detection methods for deployment type '{0}' from the XML" -f $detectionMethodXmlFile) -LogId $LogId -Severity 3
+                                Write-Host ("There was an error getting the local detection methods for deployment type '{0}' from the XML" -f $detectionMethodXmlFile) -ForegroundColor Red
+                            }
+                        }
+                    }
+                    
+                    # Add detection method details to the PSCustomObject
+                    $deploymentObject | Add-Member NoteProperty -Name DetectionTypeScriptRunAs32Bit -Value $detectionTypeScriptRunAs32Bit
+                    $deploymentObject | Add-Member NoteProperty -Name DetectionTypeScriptType -Value $detectionTypeScriptType
+                    $deploymentObject | Add-Member NoteProperty -Name DetectionTypeExecutionContext -Value $detectionTypeExecutionContext
+                    $deploymentObject | Add-Member NoteProperty -Name DetectionMethodXmlFile -Value $detectionMethodXmlFile
+                    $deploymentObject | Add-Member NoteProperty -Name DetectionMethodJsonFile -Value $detectionMethodJsonFile
+
+                    Write-Log -Message ("Application_Id = '{0}', Application_Name = '{1}', Application_LogicalName = '{2}', LogicalName = '{3}', Name = '{4}', `
+                    Technology = '{5}', ExecutionContext = '{6}', InstallContext = '{7}', InstallCommandLine = '{8}', UninstallSetting = '{9}', UninstallContent = '{10}', `
+                    UninstallCommandLine = '{11}', ExecuteTime = '{12}', MaxExecuteTime = '{13}', DetectionProvider = '{14}', DetectionTypeScriptRunAs32Bit = '{15}', `
+                    DetectionTypeScriptType = '{16}', DetectionTypeExecutionContext = '{17}', DetectionMethodXmlFile = '{18}', DetectionMethodJsonFile = '{19}'" -f `
                             $ApplicationId, `
                             $xmlContent.AppMgmtDigest.Application.title.'#text', `
                             $xmlContent.AppMgmtDigest.Application.LogicalName, `
@@ -92,10 +267,17 @@ function Get-DeploymentTypeInfo {
                             $uninstallLocation, `
                             $object.Installer.CustomData.UninstallCommandLine, `
                             $object.Installer.CustomData.ExecuteTime, `
-                            $object.Installer.CustomData.MaxExecuteTime) -LogId $LogId
+                            $object.Installer.CustomData.MaxExecuteTime, `
+                            $object.Installer.DetectAction.Provider, `
+                            $detectionTypeScriptRunAs32Bit, `
+                            $detectionTypeScriptType, `
+                            $detectionTypeExecutionContext, `
+                            $detectionMethodXmlFile, `
+                            $detectionMethodJsonFile) -LogId $LogId
 
                     # Output the deployment type object
-                    Write-Host "`n$deploymentObject`n" -ForegroundColor Green
+                    Write-Host "`nDeplopymentType Details extracted" -ForegroundColor Cyan
+                    Write-Host "$deploymentObject`n" -ForegroundColor Green
 
                     # Add the deployment type object to the array
                     $deploymentTypes += $deploymentObject          
@@ -112,7 +294,7 @@ function Get-DeploymentTypeInfo {
         catch {
             Write-Log -Message ("Could not get deployment type information for application Id '{0}'" -f $ApplicationId) -LogId $LogId -Severity 3
             Write-Warning -Message ("Could not get deployment type information for application id '{0}'" -f $ApplicationId)
-            Get-ScriptEnd -LogId $LogId -Message $_.Exception.Message
+            Get-ScriptEnd -LogId $LogId -ErrorMessage $_.Exception.Message
         }
     }
 }